requirements.txt
asgiref==3.5.2
astroid==2.12.9
asttokens==2.0.8
autoflake==1.5.3
backcall==0.2.0
backports.zoneinfo==0.2.1
certifi==2022.6.15.1
cfgv==3.3.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
dill==0.3.5.1
distlib==0.3.6
dj-database-url==1.0.0
Django==4.1.1
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.13.1
drf-yasg==1.21.3
executing==1.0.0
filelock==3.8.0
flake8==5.0.4
identify==2.5.5
idna==3.3
inflection==0.5.1
ipdb==0.13.9
ipython==8.5.0
isort==5.10.1
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mccabe==0.7.0
nodeenv==1.7.0
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.2
pre-commit==2.20.0
prompt-toolkit==3.0.31
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.9.1
pyflakes==2.5.0
Pygments==2.13.0
pylint==2.15.2
pyparsing==3.0.9
python-decouple==3.6
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
stack-data==0.5.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
traitlets==5.3.0
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.16.5
wcwidth==0.2.5
wrapt==1.14.1
